+++ geopandas/_vectorized.py
@@ -700,6 +700,16 @@ def minimum_bounding_radius(data):
         )
 
 
+def node(data):
+    if compat.USE_SHAPELY_20:
+        return shapely.node(data)
+    else:
+        raise NotImplementedError(
+            f"shapely >= 2.0 is required, "
+            f"version {shapely.__version__} is installed"
+        )
+
+
 #
 # Binary predicates
 #